ver6

+++ calcu.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int num1, num2, result;
+    float num1, num2, result;
     char operator;
 
     printf("Enter first number: ");
@@ -10,25 +10,25 @@ int main() {
     printf("Enter an operator (+, -, *, /): ");
     scanf(" %c", &operator);  
     printf("Enter second number: ");
-    scanf("%d", &num2);
+    scanf("%f", &num2);
 
     switch (operator) {
         case '+':
             result = num1 + num2;
-            printf("Result: %d + %d = %d\n", num1, num2, result);
+            printf("Result: %f + %f = %f\n", num1, num2, result);
             break;
         case '-':
             result = num1 - num2;
-            printf("Result: %d - %d = %d\n", num1, num2, result);
+            printf("Result: %f - %f = %f\n", num1, num2, result);
             break;
         case '*':
             result = num1 * num2;
-            printf("Result: %d * %d = %d\n", num1, num2, result);
+            printf("Result: %f * %f = %f\n", num1, num2, result);
             break;
         case '/':
             if (num2 != 0) {
                 result = num1 / num2;
-                printf("Result: %d / %d = %d\n", num1, num2, result);
+                printf("Result: %f / %f = %f\n", num1, num2, result);
             } else {
                 printf("Error! Division by zero is not allowed.\n");
             }
